dev-requirements.txt
#SPDX-FileCopyrightText: 2024 PNED G.I.E.
#
#SPDX-License-Identifier: Apache-2.0

pytest-ckan
